requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
google-generativeai
opencv-python-headless
pymupdf
python-dotenv
aiofiles
numpy
pillow
